+++ neoforge/changelog.md
@@ -8,3 +8,4 @@
 - Reworked microphone amplification
 - Renamed microphone amplification to microphone gain
 - Fixed French translation
+- Added audio level meters to adjust volumes section